fixed a formatting issue where numbers wouldn't line up

+++ src/constants.h
@@ -5,6 +5,6 @@
 
 const int BLOCK_SIZE = 5;
 const int SIZE = BLOCK_SIZE*BLOCK_SIZE;
-const int BLOCK_SUM = (1 << SIZE) - 1;
+const long BLOCK_SUM = (1 << SIZE) - 1;
 
 #endif
